refactor(ui-react): move CircuitBoard to own child folder
