+++ packages/ui-react/src/common/redux/reducers.js
@@ -1,10 +1,8 @@
-import { combineReducers, Reducer } from 'redux'
+import { combineReducers } from 'redux'
 
 import { mode } from '../../modules/Mode'
-import { reducer as displaySettings } from '../../modules/App/displaySettings.reducer'
 import { chips, selectedChipId } from '../../modules/Chip'
-import { reducer as selectedNodeId } from '../../modules/CircuitBoard/selectedNodeId.reducer'
-import { circuitBoardReducer as circuitBoard } from '../../modules/CircuitBoard/circuitBoard.reducer'
+import { circuitBoard, displaySettings, selectedNodeId } from '../../modules/CircuitBoard'
 
 export const rootReducer = combineReducers({
   mode,
