--- packages/ui-react/src/modules/App/displaySettings.selectors.js
@@ -1,2 +1,0 @@
-
-export const selectDisplaySettings = (state) => (state.displaySettings)
