+++ packages/ui-react/src/modules/Chip/ChipDetail/ChipDetail.container.js
@@ -1,10 +1,10 @@
 import { connect } from 'react-redux'
 
-import { ChipDetail } from './ChipDetail'
-import { selector } from './selectors'
+import { ChipDetail as ChipDetailComponent } from './ChipDetail'
+import { ChipDetailSelector } from './ChipDetails.selectors'
 
-const mapStateToProps = selector
+const mapStateToProps = ChipDetailSelector
 
 const mapDispatchToProps = null
 
-export const ChipDetail = connect (mapStateToProps, mapDispatchToProps) (ChipDetail)
+export const ChipDetail = connect (mapStateToProps, mapDispatchToProps) (ChipDetailComponent)
