+++ packages/ui-react/src/modules/Chip/ChipDetail/ChipDetails.selectors.js
@@ -2,9 +2,9 @@ import { createSelector } from '@reduxjs/toolkit'
 
 import { selectedChipSelector } from '../chips.selectors'
 
-export const selector = createSelector(
+export const ChipDetailSelector = createSelector(
   [selectedChipSelector],
-  (chip) => ({
+  chip => ({
     chip,
   }),
 )
