+++ packages/ui-react/src/modules/Chip/ChipList/ChipList.container.js
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux'
 
-import { ChipList } from './ChipList'
-import { ChipListSelector } from '../chips.selectors'
+import { ChipList as ChipListComponent } from './ChipList'
+import { ChipListSelector } from './ChipList.selectors'
 import { setSelectedChipId } from '../selectedChipId.reducer'
 
 const mapStateToProps = ChipListSelector
@@ -10,4 +10,4 @@ const mapDispatchToProps = dispatch => ({
   onSelect: chipId => dispatch (setSelectedChipId (chipId)),
 })
 
-export const ChipList = connect (mapStateToProps, mapDispatchToProps) (ChipList)
+export const ChipList = connect (mapStateToProps, mapDispatchToProps) (ChipListComponent)
