+++ packages/ui-react/src/modules/Chip/ChipList/ChipList.jsx
@@ -5,7 +5,7 @@ import styles from './ChipList.module.css'
 
 export const ChipSummary = ({ name, onClick, isSelected }) => (
   <div className={isSelected ? styles.selectedChipId : styles.chip} onClick={onClick}>
-    <text>{name}</text>
+    {name}
   </div>
 )
 
@@ -13,7 +13,7 @@ export const ChipList = ({ selectedChipId, chips, onSelect }) => (
   <div>
     <div className={styles.header}>Chips</div>
     {map (({ id, name }) => (
-      <ChipSummary name={name} onClick={() => onSelect(id)} isSelected={(selectedChipId === id)} />
+      <ChipSummary key={id} name={name} onClick={() => onSelect(id)} isSelected={(selectedChipId === id)} />
     )) (chips)}
   </div>
 )
