+++ packages/ui-react/src/modules/Chip/ChipList/ChipList.selectors.js
@@ -1,0 +1,11 @@
+import { createSelector } from '@reduxjs/toolkit'
+import { values } from 'ramda'
+import { chipsSelector, selectedChipIdSelector } from '../chips.selectors'
+
+export const ChipListSelector = createSelector (
+  [chipsSelector, selectedChipIdSelector],
+  (chips, selectedChipId) => ({
+    chips: values (chips),
+    selectedChipId,
+  }),
+)
