+++ packages/ui-react/src/modules/CircuitBoard/components.js
@@ -1,1 +1,1 @@
-export { CircuitBoard } from './CircuitBoard'
+export { CircuitBoard } from './CircuitBoard/CircuitBoard.container'
