+++ packages/ui-react/src/modules/CircuitBoard/displaySettings.reducer.js
