+++ packages/ui-react/src/modules/CircuitBoard/index.js
@@ -1,3 +1,3 @@
-import CircuitBoard from './CircuitBoard.container'
-
-export default CircuitBoard
+export { reducer as displaySettings } from './displaySettings.reducer'
+export { reducer as circuitBoard } from './circuitBoard.reducer'
+export { reducer as selectedNodeId } from './selectedNodeId.reducer'
